+++ R/multi_peak_simulation.R
@@ -1,3 +1,37 @@
+#' Simulate a multimodal histogram composed as the sum of data from sampling several distributions
+#' WARNING: Comprehensive error checking required before export
+#' TODO: potentially allow for user specification of sampled distributions (i.e. actual dist)
+#'
+#' @param N number of points to sample from the distribution
+#' @param unif_length range of uniform distribution length (vector of length 2)
+#' @param norm_sd range of normal standard deviations (vector of length 2)
+#' @param gamma_shape range of gamma shape (vector of length 2)
+#' @param eps range of hyperparameter epsilon (vector of length 2)
+#' @param noise range of noise as a proportion of total counts (vector of length 2)
+#' @param max_uniform NULL forces an equal probability of sampling TRUE or FALSE, otherwise deterministic
+#' @param remove_low_entropy NULL forces an equal probability of sampling TRUE or FALSE, otherwise deterministic
+#' @param truncated_models NULL forces an equal probability of sampling TRUE or FALSE, otherwise deterministic
+#' @param peaks range of number of peaks to sample from (vector of length 2)
+#' @param peak_shift range of number of standard deviations of the first peak to sample from (vector of length 2) - this is used to shift subsequent peaks
+#' @param metrics metrics for segment and fit
+#' @param seed numeric seed
+#' @param run_segment_and_fit logical; whether to run segment and fit or not
+#' @param include_simulated_data logical; whether to return simulated data
+#' @param include_fit_data logical; whether to return the HistogramFit object
+#'
+#' @return A list of results
+#' \describe{
+#'     \item{actual_peaks}{a data frame of parameters used for simulation, one row per simulated peak}
+#'     \item{seg_results}{if `run_segment_and_fit`, a data frame for peaks identified by segment_and_fit}
+#'     \item{overlap}{if `run_segment_and_fit`, a data frame for overlap between simulated and fitted peaks}
+#'     \item{peak_data}{if `include_simulated_data`, simulated peak counts will be returned}
+#'     \item{noise_data}{if `include_simulated_data`, simulated noise counts will be returned}
+#'     \item{hz_model}{if `include_fit_data`, the HistogramFit object will be returned}
+#'}
+#'
+#' @examples \dontrun{
+#' simulated_histogram <- random_multi_peak_sim()
+#'}
 random_multi_peak_sim <- function(
     N = c(25, 500),
     unif_length = c(6, 25),
@@ -11,9 +45,13 @@ random_multi_peak_sim <- function(
     peaks = 2:3,
     peak_shift = c(2, 8), # Peak shift in standard deviations from first peak
     metrics = c('mle', 'jaccard', 'intersection', 'ks', 'mse', 'chisq'),
-    return_fit = FALSE,
-    seed = as.integer(sub('^16', sample(1:9, size = 1), as.integer(Sys.time())))
+    seed = as.integer(sub('^16', sample(1:9, size = 1), as.integer(Sys.time()))),
+    run_segment_and_fit = TRUE,
+    include_simulated_data = FALSE,
+    include_fit_data = FALSE
   ) {
+
+  set.seed(seed);
   if (is.null(max_uniform)) max_uniform <- sample(c(TRUE, FALSE), size = 1)
   if (is.null(remove_low_entropy)) remove_low_entropy <- sample(c(TRUE, FALSE), size = 1)
   if (is.null(truncated_models)) truncated_models <- sample(c(TRUE, FALSE), size = 1)
@@ -21,7 +59,6 @@ random_multi_peak_sim <- function(
   .sample_unif <- function(x, n = 1) {
     runif(n, min = x[1], max = x[2])
   }
-  set.seed(seed);
 
   if (length(peaks) > 1) {
     peaks_sim <- sample(x = peaks, size = 1)
@@ -29,12 +66,14 @@ random_multi_peak_sim <- function(
     peaks_sim <- peaks
     }
 
+  set.seed(seed);
   sim_dist <- sample(c('norm', 'unif', 'gamma'), size = peaks_sim, replace = TRUE)
   N_sim <- round(.sample_unif(N, n = peaks_sim))
   eps_sample <- .sample_unif(eps, n = 1)
   noise_sim <- .sample_unif(noise)
   N_noise_sim <- round(sum(N_sim) * noise_sim)
 
+  set.seed(seed);
   multi_histogram_data <- lapply(seq_len(peaks_sim), function(i) {
     # Mean is 0 for each of the distributions
     if (sim_dist[[i]] == 'norm') {
@@ -70,8 +109,7 @@ random_multi_peak_sim <- function(
     })
 
   peak_shift_sample <- 0
-  # Shift each eap
-  # In standard deviations
+  # Shift each peak (a sampled number of standard deviations)
   if (peaks_sim > 1) {
     peak_shift_sample <- cumsum(c(0, .sample_unif(peak_shift, n = peaks_sim - 1)))
     peaK_shift_actual <- multi_histogram_data[[1]]$sd_dist * peak_shift_sample
@@ -120,8 +158,11 @@ random_multi_peak_sim <- function(
   peak_min <- unlist(lapply(peak_histograms, function(x) head(x$interval_start, n = 1)))
   peak_max <- unlist(lapply(peak_histograms, function(x) tail(x$interval_end, n = 1)))
 
+  optima <- find_local_optima(merged_peak_histograms_noise)
+
     res <- list(
         N = N_sim,
+        local_optima = sum(sapply(optima, length)),
         param = unlist(lapply(multi_histogram_data, function(x) x$param)),
         peak_num = seq_along(peak_min),
         peak_min = peak_min,
@@ -141,21 +182,24 @@ random_multi_peak_sim <- function(
         peaks = peaks_sim
         )
 
-    print.res <- as.data.frame(res)
-    print('PARAMS: ')
-    print(print.res)
+    # long format, one peak per line
+    actual_peaks <- as.data.frame(res)
 
-  timing <- system.time({
-    seg_results_mod <- try({
-      segment_and_fit(
-        merged_peak_histograms_noise,
-        max_uniform = max_uniform,
-        remove_low_entropy = remove_low_entropy,
-        metric = metrics,
-        eps = eps_sample,
-        truncated_models = truncated_models,
-        metric_weights = sqrt(rev(seq_along(metrics))),
-        seed = seed
+    print('PARAMS: ')
+    print(actual_peaks)
+
+  if (run_segment_and_fit) {
+    timing <- system.time({
+      seg_results_mod <- try({
+        segment_and_fit(
+          merged_peak_histograms_noise,
+          max_uniform = max_uniform,
+          remove_low_entropy = remove_low_entropy,
+          metric = metrics,
+          eps = eps_sample,
+          truncated_models = truncated_models,
+          metric_weights = sqrt(rev(seq_along(metrics))),
+          seed = seed
         )
       }, silent = TRUE)
     peak_start <- seg_results_mod$interval_start[seg_results_mod$p[, 'start']]
@@ -163,31 +207,15 @@ random_multi_peak_sim <- function(
     seg_results <- summarize_results_error(seg_results_mod)
     })
 
-  seg_results <- cbind.data.frame(
-    seg_results,
-    t(unclass(timing))[, c('user.self', 'sys.self', 'elapsed'), drop = FALSE]
+    seg_results <- cbind.data.frame(
+      seg_results,
+      t(unclass(timing))[, c('user.self', 'sys.self', 'elapsed'), drop = FALSE]
     )
-
-    # long format, one peak per line
-    actual_peaks <- data.frame(res)
 
     # Joined by seed
     seg_results$seed <- seed
 
     # Compute overlap data for peaks
-    overlap_data <- apply(actual_peaks[, c('peak_min', 'peak_max')], 1, function(x) {
-      which.max(
-        sapply(seq_along(peak_start), function(i) {
-          overlap_size(
-            a1 = x['peak_min'],
-            a2 = x['peak_max'],
-            b1 = peak_start[i],
-            b2 = peak_end[i]
-            )
-          }, simplify = FALSE)
-        )
-      }, simplify = FALSE)
-
     overlap_data <- merge(
       actual_peaks[, c('peak_num', 'peak_min', 'peak_max')],
       data.frame(
@@ -219,16 +247,40 @@ random_multi_peak_sim <- function(
       truncated_models = truncated_models,
       seed = seed
       )
-
-    rtn <- list(
-        actual_peaks = actual_peaks,
-        seg_results = seg_results,
-        overlap = overlap_data
-        )
-
-    if (return_fit) {
-      rtn$hz_fit <- seg_results_mod
-      }
-
-    return(rtn)
   }
+
+  rtn <- list(
+    "actual_peaks" = actual_peaks
+  )
+
+  if(run_segment_and_fit){
+    rtn <- c(
+      rtn,
+      list(
+        "seg_results" = seg_results,
+        "overlap" = overlap_data
+      )
+    )
+  }
+
+  if(include_simulated_data){
+    rtn <- c(
+      rtn,
+      list(
+        "peak_data" = merged_raw_data,
+        "noise_data" = noise_data
+      )
+    )
+  }
+
+  if(include_fit_data && run_segment_and_fit){
+    rtn <- c(
+      rtn,
+      list(
+        "hz_model" = seg_results_mod
+      )
+    )
+  }
+
+  return(rtn)
+}